NO-ISSUE: Fix AppProject setup docs
This PR updates the README for the sample git setup in the docs to correct the hub cluster preparation. It currently says to add the ClusterTemplate resource to the `spec.clusterResourceWhitelist` field of the AppProject even though the ClusterTemplate is a namespace scoped resource. Instead, it should go under the `spec.namespaceResourceWhitelist` field.

Signed-off-by: Kirsten Laskoski <klaskosk@redhat.com>